save code for v3 of training data with missing voxel check

+++ make_flashmatch_training_data.py
@@ -1,5 +1,5 @@
 from __future__ import print_function
-import os,sys,argparse
+import os,sys,argparse,signal
 from math import fabs
 
 parser = argparse.ArgumentParser(description='Make MC flashmatch training data from ROOT file. Store into petastorm.')
@@ -9,6 +9,9 @@ parser.add_argument('-mc',"--in-mcinfo",required=True,type=str,help="path to mci
 parser.add_argument('-op',"--in-opreco",required=True,type=str,help="path to opreco root file")
 parser.add_argument('-v',"--verbosity",type=int,default=0,help='Set Verbosity Level [0=quiet, 2=debug]')
 parser.add_argument('-e',"--entry",type=int,default=None,help='Run specific entry')
+parser.add_argument('-n',"--num-entries",type=int,default=None,help='Run n entries')
+parser.add_argument('-xw',"--no-write",default=False,action='store_true',help="If flag given, we will not write to DB. For debugging.")
+parser.add_argument('-ow',"--over-write",default=False,action='store_true',help="If flag given, will overwrite existing database chunk without user check")
 args = parser.parse_args(sys.argv[1:])
 
 import ROOT as rt
@@ -36,13 +39,9 @@ test script that demos the Flash Matcher class.
 
 ### DEV OUTPUTS
 output_url="file:///"+args.db_folder
-WRITE_TO_SPARK = True # For debug. Turn off to avoid modifying database
-
-start_entry = 0
-end_entry = -1
-if args.entry is not None:
-    start_entry = args.entry
-    end_entry = start_entry+1
+WRITE_TO_SPARK = not args.no_write # For debug. Turn off to avoid modifying database
+ONLY_ANODE_CATHODE = False
+OVERWRITE_INPUT_TIMEOUT_SECS = 10
 
 sourcefile = os.path.basename(args.in_mcinfo)
 input_larcv_rootfile_v = [args.in_larcvtruth]
@@ -123,18 +122,58 @@ nentries = iolcv.get_n_entries()
 print("Number of entries (LARCV): ",nentries)
 print("Number of entries (LARLITE): ",ioll.get_entries())
 
+start_entry = 0
+end_entry = nentries
+if args.entry is not None:
+    start_entry = args.entry
+    end_entry = start_entry+1
+if args.num_entries is not None:
+    if end_entry<0:
+        end_entry = nentries
+    else:
+        end_entry = start_entry + args.num_entries
+        if end_entry>nentries:
+            end_entry = nentries
+
+print("RUN FROM ENTRY ",start_entry," to ",end_entry)
+
+
 if WRITE_TO_SPARK:
+    print("********** WRITING TO SPARK DB ***************")
     spark_session = SparkSession.builder.config('spark.driver.memory', '2g').master('local[2]').getOrCreate()
     sc = spark_session.sparkContext
 
     # remove past chunk from database
     chunk_folder = args.db_folder+"/\'sourcefile=%s\'"%(sourcefile)
     if os.path.exists(chunk_folder):
-        print("remove old database folder")
-        os.system("rm -r %s"%(chunk_folder))
+        print("*****  removing old database chunk folder: ",chunk_folder,"  **********")
+        if args.over_write:
+            print("proceeding without check due to '--over-write' flag has been given")
+            os.system("rm -r %s"%(chunk_folder))
+        else:
+            print("[enter 'y' or 'Y' to allow overwrite. any other input stops program.] (to over-write without check, give '--over-write')")
+            # defining a handler
+            def handle_no_input(signum,frame):
+                raise IOError("user input check timed out")
+            signal.signal(signal.SIGALRM, handle_no_input)
+            
+            signal.alarm(OVERWRITE_INPUT_TIMEOUT_SECS)
+            userinput = 'n'
+            try:
+                userinput = input()
+            except:
+                print("over-write check timed out. stopping program.")
+                sys.exit(1)
+            signal.alarm(0) # disable alarm after success
+            # should be a check here unless overwrite argument given
+            if userinput in ['y','Y']:
+                print("user OK( entered: %s ) provided to delete chunk folder"%(userinput))
+                os.system("rm -r %s"%(chunk_folder))
+            else:
+                print("overwrite not allowed by user input: ",userinput)
+                sys.exit(0)
+        
 
-if end_entry<0:
-    end_entry = nentries
 
 row_data = []
     
@@ -153,10 +192,54 @@ for ientry in range( start_entry, end_entry ):
     subrun  = ioll.subrun_id()
     eventid = ioll.event_id()
 
-    truth_correct_tdrift = True
+    # the function below uses the wire plane images in stored in the larcv io manager (iolcv)
+    # and proposes possible 3d spacepoints consistent with the pattern of ionization in the images
+    # it then uses the 'truth' information in the simulation to label those proposed spacepoints
+    # Labels relevant here are:
+    #  1. is the spacepoint real or a 'ghost' point?
+    #  2. the ID number from Geant4 associated to the particle that
+    #       left the ionization deposite the spacepoint represents
+    #  3. the ID number of the "ancestor" particle that ultimately led to the current particle
+    #  4. The value of the pixels on each plane associated to each point.
+    # After labels for all the points are made, we define voxels over a 3D rectangle and assign
+    #     points to individual voxels in the grid.
+    #     (A voxel may be assigned more than one point.)
+    #     (each point is assigned to only one voxel)
+    # Transferring labels to the voxels requires some may to reduce the set of labels from the points.
+    # For a given voxel, with a set of points
+    #  1. if any point has a 'true' label, the voxel is labeled as 'true'
+    #  2. for our purposes, we will not need to reduce track id labels. we will need a set of voxels    
+    #     to store the ionization deposited by one 'interaction cascade'. A cascade is paired
+    #     to a reconstructed optical flash in order to train our network. When we select the set of
+    #     voxels to represent the ionization, we can take all voxels associated to points carrying
+    #     the ancestor ID of the interaction cascade.
+    # One complication is that particles coming from a neutrino interaction have different
+    #  ancestor IDs because they are given as 'primary' particles to Geant4. We change
+    #  ancestor IDs for particles coming from neutrino interactions to a common ID of '0'.
+    #
+    # Note: if 'truth_correct_tdrift' is TRUE, then we will use the truth information to
+    #   remove the t0-offset for each point.
+    # We can do this, beacause a true point (usually) is labeled by a track ID.
+    # We can then look up the true time the particle crossed the detector and subtract the 
+    #   the x-position offset by (t_particle - t_trigger)*drift_velocity.
+    #   (We use units of usec and cms).
+    # We make this position adjustment to points BEFORE assigning them to voxels.
+    # The result is that all voxels should be within the voxels corresponding to the drift volume
+    #   in time with the beam.
+    truth_correct_tdrift = True    
     voxelizer.process_fullchain_withtruth( iolcv, ioll, adc_name, adc_name, truth_correct_tdrift )
+
+    # for debug: dump out the particle tree graph, showing the particles whose information we have.
+    #mcpg = ublarcvapp.mctools.MCPixelPGraph()
+    #mcpg.buildgraphonly( ioll )
+    #mcpg.printGraph(0,False)
     
-    # match reco flashes to true track and shower information
+    # match reco flashes to true track and shower information.
+    # this version uses the voxel information as well to
+    # check if examples are not missing a significant fraction of
+    # voxels with charge.
+    # Note: the process function needs to know if the voxels already have their
+    # true t0 shift removed!!!
     print("Run opdataprep")    
     fmutil.process( ioll, voxelizer )
     fmutil.printMatches()
@@ -173,21 +256,18 @@ for ientry in range( start_entry, end_entry ):
 
     naccepted = 0
 
+    # loop through the truth-matched reconstruction optical flashes
+    # that have been matched to a set of particles in the saved simulation info.
     for iflash in range( fmutil.recoflash_v.size() ):
 
         flash = fmutil.recoflash_v.at(iflash)
     
-        # get flash match vectors
-        #coord_v = std.vector("std::vector<int>")()
-        #feat_v  = std.vector("std::vector<float>")()
-        #opdataprep.getChargeVoxelsForFlash( flash, voxelizer, coord_v, feat_v )
-
-        # get the right flash pe vector
-        #if flash.producerid>=0:
-        #    flash_np = flash_np_v[ (flash.producerid,flash.index) ]
-        #else:
-        #    flash_np = np.zeros( 32, dtype=np.float32 )            
-        
+        # the following function provides us with the info we need.
+        # below returns a dictionary with keys:
+        # 'flashpe':   (1,32) float32 array with total PE in light pulse seen by each PMT
+        # 'voxcoord':  (N,3) int64 array with the index of the N voxels that contain space points
+        # 'voxcharge': (N,3) float32 array with the charge sum of points for each plane.
+        #              For each plane, an individual pixel is only able to contribute once to the total.
         data_dict = fmutil.make_opmodel_data_dict( flash, voxelizer, ioll )
         
         print("flash[",iflash,"]")
@@ -198,72 +278,52 @@ for ientry in range( start_entry, end_entry ):
         print("  coord x-bounds: [",data_dict['voxcoord'][:,0].min(),",",data_dict['voxcoord'][:,0].max(),"]")
         print("  coord y-bounds: [",data_dict['voxcoord'][:,1].min(),",",data_dict['voxcoord'][:,1].max(),"]")
         print("  coord z-bounds: [",data_dict['voxcoord'][:,2].min(),",",data_dict['voxcoord'][:,2].max(),"]")        
-        print("  frac of track traj. in tpc with voxel charge: ",fmutil.flash_track_frac_intpc_w_charge.at(iflash))            
+        print("  frac of track traj. in tpc with voxel charge: ",fmutil.flash_track_frac_intpc_w_charge.at(iflash))
 
-        # we have to use the true t0 time and remove the shift
-        # shift should occur when making voxels
-        t0shift_cm = (flash.tick-3200.0)*0.5*driftv
-        t0shift_vox = int(t0shift_cm/voxel_len)
+        # get the t_drift adjustment.
+        # emperical way is to use anode or cathode crossers only
+        # or we can use the truth
 
-        # we are missing the drift time
-        # thus, can only know true position in data for training the model
-        # if when we subtract the t0shift
-        # the end of the tracks are at the anode or cathode
-        vox_xmin = data_dict['voxcoord'][:,0].min()
-        vox_xmax = data_dict['voxcoord'][:,0].max()
-
-        # remove t0shift
-        print("  t0shift_cm: ",t0shift_cm)
-        print("  t0shift_vox: ",t0shift_vox)
-
-        # is xmin-t0shift close to zero?
-        # is xmax-t0shift close to 256?
-        anode_dt = vox_xmin-t0shift_vox
-        cathode_dt = (vox_xmax-(t0shift_vox+256.0/voxel_len))
-        print("  anode_dt: ",anode_dt)
-        print("  cathode_dt: ",cathode_dt)
-
-        keep = False
-        if abs(anode_dt)<=2:
-            win_xmin = vox_xmin
-            win_xmax = win_xmin + int(260.0/voxel_len)
-            keep = True
-            print("  detected as anode-crossing")
-        elif abs(cathode_dt)<=4:
-            win_xmax = vox_xmax
-            win_xmin = vox_xmax - int(260.0/voxel_len)
-            keep = True
-            print("  detected as cathode-crossing")
-        elif flash.producerid==0:
-            print("cheating: saving neutrinos")
-            keep = True
-            win_xmin = index_tpc_origin[0]
-            win_xmax = index_tpc_end[0]
-
-        if not keep:
-            print("  neither anode or cathode crossing")
+        # skip null flash outliers for now
+        if flash.producerid==-1:
             continue
-
+        
+        # since we already removed the t0shift, the voxels should be in the right place
+        # just crop in the main drift window
+        win_xmin = index_tpc_origin[0] 
+        win_xmax = index_tpc_end[0]
         print("  shifted tpc bounds in x: [",win_xmin,",",win_xmax,"]")
         
         # crop around tpc
+        # below gives me a T value for each voxel, that falls inside an interval for a given dimension.
         voxcoord_above_xbound = (data_dict['voxcoord'][:,0]>=win_xmin)*(data_dict['voxcoord'][:,0]<=win_xmax)
         voxcoord_above_ybound = (data_dict['voxcoord'][:,1]>=index_tpc_origin[1])*(data_dict['voxcoord'][:,1]<=index_tpc_end[1]+1)
         voxcoord_above_zbound = (data_dict['voxcoord'][:,2]>=index_tpc_origin[2])*(data_dict['voxcoord'][:,2]<=index_tpc_end[2]+1)
+        # create a box interval by requiring that voxels fall within the intervals of all the dimensions
         intpc = voxcoord_above_xbound*voxcoord_above_ybound*voxcoord_above_zbound
+        # select voxels falling within my 3D box interval representing the TPC
         voxcoord_intpc = data_dict['voxcoord'][intpc[:],:]
         voxfeat_intpc  = data_dict['voxcharge'][intpc[:],:]
-        voxcoord_intpc[:,0] += t0shift_vox
-        print("  shifted the xindex: ",voxcoord_intpc[:,0].min(),",",voxcoord_intpc[:,0].max())
-        print("  voxcoord IN TPC: ",voxcoord_intpc.shape," ",voxcoord_intpc.dtype)
-        print("  voxfeat IN TPC: ",voxfeat_intpc.shape," ",voxfeat_intpc.dtype)
+        # subtract off the index value representing the lower-x corner of the TPC
+        # this effectively shrinks the voxel set to only those inside the TPC
+        voxcoord_intpc[:,0] -= win_xmin
+        print("  x-index bounds after mask and shifted: ",voxcoord_intpc[:,0].min(),",",voxcoord_intpc[:,0].max())
+        print("  voxcoord IN TPC shape and type: ",voxcoord_intpc.shape," ",voxcoord_intpc.dtype)
+        print("  voxfeat IN TPC shape and type: ",voxfeat_intpc.shape," ",voxfeat_intpc.dtype)
+
+        #print("voxcoord_intpc ===================")
+        #print(voxcoord_intpc)
+        #print("voxcoord_intpc ===================")        
+
+        # if no more voxels remaining, just skip this example.
+        if ( voxcoord_intpc.shape[0]<1 ):
+            continue
         
-        
-        # make the row of data
+        # make the row of data to save in our data
         row = {"sourcefile":sourcefile,
                "run":run,
                "subrun":subrun,
-               "event":eventid,               
+               "event":ientry,               
                "matchindex":int(naccepted),               
                "flashpe":data_dict['flashpe'],
                "coord":voxcoord_intpc,
@@ -271,8 +331,9 @@ for ientry in range( start_entry, end_entry ):
                "ancestorid":int(flash.ancestorid)}
         naccepted += 1
         row_data.append( dict_to_spark_row(FlashMatchSchema,row) )
-
-    print("number of training examples made: ",naccepted)
+        
+    # end of loop over flash-voxel pairs
+    print("number of training examples made in this event: ",naccepted)
 
 #end of event loop
 
@@ -291,5 +352,7 @@ if WRITE_TO_SPARK:
                      .mode(write_mode) \
                      .parquet( output_url )
         print("spark write operation")
+else:
+    print("Skipping writing to data base")
         
 print("=== FIN ==")
